+++ src/components/post/PostCard/style.ts
